requirements.txt
Django==3.0.7
django-crispy-forms==1.9.1
djangorestframework==3.11.0
facebook-sdk==3.1.0
